backend/requirement.txt
altgraph==0.17.4
asgiref==3.8.1
Django==5.0.8
django-cors-headers==4.4.0
djangorestframework==3.15.2
GDAL==3.9.2
macholib==1.16.3
packaging==24.1
psycopg2-binary==2.9.9
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
setuptools==73.0.1
sqlparse==0.5.1
